#77 Move migrations into own directory
